Code updated: . Removed all unwanted code lines . Separated the UICode and Animation code into new files
