+++ Freyja/Reusables/View/KnobView/KnobMotorViewModel.swift
@@ -86,8 +86,7 @@ extension KnobMotorViewModel: KnobViewToViewModelProtocol {
         guard let angle = getAngle(from: location) else { return }
         rotationStartedAngle = angle
         switch mode {
-        case .knob:
-//            view?.cancelAllAnimations()
+        case .knob: 
             break
         case .slingShot:
             view?.cancelAllAnimations()
